block control word indexing fix
This fixes a similar parsing on invalid block control word issue that we saw previously. The 'offset' variable keeps track of where on a given record the loop is and a block must be at least 2 words long. When at the final (512th) index of the record and attempting to process the next block we must first go to the next record in the timeseries. See line 307.

However, we've used python indexing for the offset variable which starts at 0. The last index of record should therefor 511 and not the 512.

Reference: https://github.com/LimnoTech/HSPsquared/issues/21#issuecomment-783755330