Fix difference in timeseries index
HSP2 looks to be running for an extra time step. To compare the results, the index of the HSPF (hbn file) and HSP2 (h5 file) timeseries needs to match. Added temporary logic to drop extra values in the HSP2 results. Longer term still solution needed and should be addressed in model code.