Fix to saveall=false bug
Reference #77. This commit modifies the save_timeseries function to generate a list of the columns in the dataframe that should be saved and now passes the dataframe to the IOManager. The logic for selectively removing timeseries prior to export is now handled by the IOManager, however the IOManager caches the dataframe will all columns in memory prior to selectively dropping columns. This allows subsequent modules to access those timeseries without saving them to the output file.